requirements.txt
streamlit==1.31.0
openai>=1.0.0
anthropic>=0.18.0
scikit-learn>=1.4.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
plotly>=5.18.0
joblib>=1.3.0
